factored out activation literal scope/pool machinery and refactored bmc to use them
