+++ patronus/src/mc.rs
@@ -2,6 +2,7 @@
 // released under BSD 3-Clause License
 // author: Kevin Laeufer <laeufer@berkeley.edu>
 
+mod act_lit;
 mod bmc;
 mod encoding;
 mod pdr;
