+++ patronus/src/mc/bmc.rs
@@ -5,6 +5,7 @@
 
 use crate::expr::*;
 use crate::mc::Witness;
+use crate::mc::act_lit::{ActLitPool, with_act_scope};
 use crate::mc::encoding::{Step, TransitionSystemEncoding, UnrollSmtEncoding};
 use crate::mc::types::{InitValue, ModelCheckResult, Result};
 use crate::mc::utils::{check_assuming, check_assuming_end, get_smt_value};
@@ -38,7 +39,7 @@ pub fn bmc(
     let constraints = sys.constraints.clone();
     let bad_states = sys.bad_states.clone();
 
-    let mut next_proxy_id = 0u64;
+    let mut act_lit_pool = ActLitPool::default();
 
     if k_max > 0 && sys.states.is_empty() {
         println!(
@@ -80,32 +81,30 @@ pub fn bmc(
                 check_assuming_end(smt_ctx)?;
             }
         } else {
-            let all_bads = bad_states
-                .iter()
-                .map(|expr_ref| enc.get_signal_at(ctx, *expr_ref, k))
-                .collect::<Vec<_>>();
+            let res = with_act_scope(ctx, smt_ctx, |ctx, smt_ctx, scope| {
+                let all_bads = bad_states
+                    .iter()
+                    .map(|expr_ref| enc.get_signal_at(ctx, *expr_ref, k))
+                    .collect::<Vec<_>>();
 
-            let bad_proxy = ctx.bv_symbol(format!("__bmc_act_{next_proxy_id}").as_str(), 1);
-            smt_ctx.declare_const(ctx, bad_proxy)?;
+                let any_bad = all_bads.into_iter().reduce(|a, b| ctx.or(a, b)).unwrap();
+                let bad_proxy = act_lit_pool.imply(ctx, smt_ctx, scope, any_bad)?;
 
-            let any_bad = all_bads.into_iter().reduce(|a, b| ctx.or(a, b)).unwrap();
-            let imp = ctx.implies(bad_proxy, any_bad);
-            smt_ctx.assert(ctx, imp)?;
-            next_proxy_id += 1;
+                let res = check_assuming(ctx, smt_ctx, [bad_proxy])?;
 
-            let res = check_assuming(ctx, smt_ctx, [bad_proxy])?;
+                // count expression uses
+                let use_counts = count_system_expr_uses(ctx, sys);
+                if res == CheckSatResponse::Sat {
+                    let wit = get_witness(sys, ctx, &use_counts, smt_ctx, &enc, k, &bad_states)?;
+                    return Ok(Some(ModelCheckResult::Fail(wit)));
+                }
+                check_assuming_end(smt_ctx)?;
+                Ok(None)
+            })?;
 
-            // count expression uses
-            let use_counts = count_system_expr_uses(ctx, sys);
-            if res == CheckSatResponse::Sat {
-                let wit = get_witness(sys, ctx, &use_counts, smt_ctx, &enc, k, &bad_states)?;
-                return Ok(ModelCheckResult::Fail(wit));
+            if let Some(fail) = res {
+                return Ok(fail);
             }
-            check_assuming_end(smt_ctx)?;
-
-            // Clean up
-            let neg_proxy = ctx.not(bad_proxy);
-            smt_ctx.assert(ctx, neg_proxy)?;
         }
 
         // advance
