+++ patronus/src/mc/pdr.rs
@@ -3,6 +3,7 @@
 // author: Michael Zhang <mxz6@cornell.edu>, Kevin Laeufer <laeufer@cornell.edu>
 
 use crate::expr::*;
+use crate::mc::act_lit::{ActLitPool, ActLitScope, with_act_scope};
 use crate::mc::bmc::start_bmc_or_pdr;
 use crate::mc::encoding::{Step, TransitionSystemEncoding};
 use crate::mc::{
@@ -24,9 +25,6 @@ const MAX_FRAMES: usize = 1000;
 
 /// Activation literal prefix
 const ACT_LIT_PREFIX: &str = "__pdr_act_";
-
-/// Activation literal prefix for frames
-const FRAME_LIT_PREFIX: &str = "__pdr_frame_act_";
 
 // -------------------------------------------------------------------------------------------------
 // Core PDR data structures
@@ -326,121 +324,6 @@ impl<E: TransitionSystemEncoding> PdrEncodingWrapper<E> {
         // Add final stepped expression to cache
         self.expr_cache.insert((expr, step), stepped);
         stepped
-    }
-}
-
-// -------------------------------------------------------------------------------------------------
-// Activation Literal Pool
-// -------------------------------------------------------------------------------------------------
-
-/// Collection of activation literals used in scope
-#[derive(Default)]
-struct ActLitScope {
-    act_lits: Vec<ExprRef>,
-}
-
-impl ActLitScope {
-    /// Permanently disable all activation literals in this scope
-    fn release(&mut self, ctx: &mut Context, smt_ctx: &mut impl SolverContext) -> Result<()> {
-        for lit in self.act_lits.drain(..) {
-            let neg_lit = ctx.not(lit);
-            smt_ctx.assert(ctx, neg_lit)?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Default)]
-struct ActLitPool {
-    /// Activation literal ID tracker
-    next_act_id: u64,
-
-    /// Cache mapping from stepped literal to activation literal
-    step_lit_cache: FxHashMap<ExprRef, ExprRef>,
-}
-
-impl ActLitPool {
-    /// Create a new activation literal
-    fn create(&mut self, ctx: &mut Context, smt_ctx: &mut impl SolverContext) -> Result<ExprRef> {
-        // Intern activation literal and define in solver
-        let lit = ctx.bv_symbol(format!("{ACT_LIT_PREFIX}{}", self.next_act_id).as_str(), 1);
-        smt_ctx.declare_const(ctx, lit)?;
-
-        // Update activation literal counter and return
-        self.next_act_id += 1;
-        Ok(lit)
-    }
-
-    /// Create a temporary activation literal that is coupled with `body` (i.e. `act => body`)
-    /// and registered into `scope`
-    fn imply(
-        &mut self,
-        ctx: &mut Context,
-        smt_ctx: &mut impl SolverContext,
-        scope: &mut ActLitScope,
-        body: ExprRef,
-    ) -> Result<ExprRef> {
-        // If solver supports compound expression assumptions, then activation literals
-        // are not needed
-        if smt_ctx.supports_check_assuming_exprs() {
-            return Ok(body);
-        }
-
-        // Create `act => body` and assert in solver
-        let act = self.create(ctx, smt_ctx)?;
-        let imp = ctx.implies(act, body);
-        smt_ctx.assert(ctx, imp)?;
-
-        // Register activation literal in scope
-        scope.act_lits.push(act);
-        Ok(act)
-    }
-
-    /// Create an activation literal for a stepped cube literal, caching the activation literal
-    /// with the associated stepped cube literal
-    ///
-    /// # Precondition
-    /// `stepped_lit` must be stepped
-    ///
-    /// # Note
-    /// Produced activation literal remains in global solver context. This is sound since
-    /// `act => stepped_lit` is a permanent fact.
-    fn step_lit_act(
-        &mut self,
-        ctx: &mut Context,
-        smt_ctx: &mut impl SolverContext,
-        stepped_lit: ExprRef,
-    ) -> Result<ExprRef> {
-        // Check cache for activation literal
-        if let Some(act) = self.step_lit_cache.get(&stepped_lit) {
-            return Ok(*act);
-        }
-
-        // Create `act => stepped_lit` and assert in solver
-        let act = self.create(ctx, smt_ctx)?;
-        let imp = ctx.implies(act, stepped_lit);
-        smt_ctx.assert(ctx, imp)?;
-
-        // Register stepped literal and its activation literal in cache
-        self.step_lit_cache.insert(stepped_lit, act);
-        Ok(act)
-    }
-}
-
-/// Execute closure with a fresh [`ActLitScope`] and clean up all used activation literals in the end
-fn with_act_scope<S: SolverContext, T>(
-    ctx: &mut Context,
-    smt_ctx: &mut S,
-    f: impl FnOnce(&mut Context, &mut S, &mut ActLitScope) -> Result<T>,
-) -> Result<T> {
-    // Create new activation literal scope, run closure, and clean up used activation literals
-    let mut scope = ActLitScope::default();
-    let res = f(ctx, smt_ctx, &mut scope);
-    let cleanup = scope.release(ctx, smt_ctx);
-
-    match res {
-        Ok(v) => cleanup.map(|()| v), // Return result from closure, noting cleanup errors
-        Err(e) => Err(e),             // Return error produced by closure
     }
 }
 
@@ -962,7 +845,7 @@ impl BasePdr {
     fn add_frame(&mut self, ctx: &mut Context, smt_ctx: &mut impl SolverContext) -> Result<()> {
         // Create new activation literal for frame
         let act = ctx.bv_symbol(
-            format!("{FRAME_LIT_PREFIX}{}", self.frames.len() + 1).as_str(),
+            format!("{ACT_LIT_PREFIX}frame_{}", self.frames.len() + 1).as_str(),
             1,
         );
         smt_ctx.declare_const(ctx, act)?;
@@ -1235,9 +1118,6 @@ pub fn pdr(
             )? {
                 // Reset solver
                 smt_ctx.restart()?;
-
-                // Clear the stepped literal cache
-                state.pool.step_lit_cache.clear();
 
                 // Cube could not be blocked: construct witness from BMC instance and fail
                 let ModelCheckResult::Fail(wit) =
